refactor(firestore): extract single write execution
